requirements.txt
Django==1.10.1
django-stdimage==2.3.3
Pillow==3.3.1
progressbar2==3.10.1
psycopg2==2.6.2
python-utils==2.0.0
six==1.10.0
